+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.9.0 (tbd)
 
+- Fixed platform providers for `piral-ng` with Angular 20 in non-standalone form (#764)
 - Improved output when installation of packages fails (#759)
 - Improved error message when template resolution fails (#763)
 - Updated `dets` to latest release